ms-product/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
click==8.1.8
dataclasses==0.6
dnspython==2.7.0
exceptiongroup==1.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.4
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.4
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
